Refactors the Diffie-Helman Finite Group classes into dedicated namespace and folder with some class renames
